refactor: migrate more virtualModule to rsbuild-virtual-module
