+++ packages/core/src/node/runtimeModule/index.ts
@@ -1,24 +1,10 @@
 import type { RsbuildPlugin } from '@rsbuild/core';
-import type { UserConfig } from '@rspress/shared';
 import { RspackVirtualModulePlugin } from 'rspack-plugin-virtual-module';
-import type { PluginDriver } from '../PluginDriver';
-import type { RouteService } from '../route/RouteService';
-import { globalStylesVMPlugin } from './globalStyles';
-import { globalUIComponentsVMPlugin } from './globalUIComponents';
 import { routeVMPlugin } from './routeData';
-import { searchHookVMPlugin } from './searchHooks';
 import { siteDataVMPlugin } from './siteData/index';
+import type { FactoryContext } from './types';
 
-export interface FactoryContext {
-  userDocRoot: string;
-  config: UserConfig;
-  isSSR: boolean;
-  runtimeTempDir: string;
-  alias: Record<string, string | string[]>;
-  routeService: RouteService;
-  pluginDriver: PluginDriver;
-}
-
+// FIXME: migrate to rsbuild-plugin-virtual-module
 type RuntimeModuleFactory = (
   context: FactoryContext,
 ) => Record<string, string> | Promise<Record<string, string>>;
@@ -33,18 +19,6 @@ export const runtimeModuleFactory: RuntimeModuleFactory[] = [
    * Also responsible for automatically importing prism languages
    */
   siteDataVMPlugin,
-  /**
-   * Generate global components from config and plugins
-   */
-  globalUIComponentsVMPlugin,
-  /**
-   * Generate global styles from config and plugins
-   */
-  globalStylesVMPlugin,
-  /**
-   * Generate search hook module
-   */
-  searchHookVMPlugin,
 ];
 
 // We will use this plugin to generate runtime module in browser, which is important to ensure the client have access to some compile-time data
@@ -91,26 +65,3 @@ export function rsbuildPluginDocVM(
     },
   };
 }
-
-export enum RuntimeModuleID {
-  GlobalStyles = 'virtual-global-styles',
-  GlobalComponents = 'virtual-global-components',
-  RouteForClient = 'virtual-routes',
-  RouteForSSR = 'virtual-routes-ssr',
-  SiteData = 'virtual-site-data',
-  SearchIndexHash = 'virtual-search-index-hash',
-  I18nText = 'virtual-i18n-text',
-  SearchHooks = 'virtual-search-hooks',
-  PrismLanguages = 'virtual-prism-languages',
-}
-
-export const runtimeModuleIDs = [
-  RuntimeModuleID.GlobalStyles,
-  RuntimeModuleID.GlobalComponents,
-  RuntimeModuleID.RouteForClient,
-  RuntimeModuleID.RouteForSSR,
-  RuntimeModuleID.SiteData,
-  RuntimeModuleID.SearchIndexHash,
-  RuntimeModuleID.I18nText,
-  RuntimeModuleID.SearchHooks,
-];
