+++ packages/core/src/node/runtimeModule/siteData/index.ts
@@ -4,9 +4,9 @@ import fs from 'node:fs/promises';
 import path from 'node:path';
 import { SEARCH_INDEX_NAME, type SiteData } from '@rspress/shared';
 import { groupBy } from 'lodash-es';
-import { type FactoryContext, RuntimeModuleID } from '..';
 import { TEMP_DIR, isProduction } from '../../constants';
 import { createHash } from '../../utils';
+import { type FactoryContext, RuntimeModuleID } from '../types';
 import { extractPageData } from './extractPageData';
 import { handleHighlightLanguages } from './highlightLanguages';
 import { normalizeThemeConfig } from './normalizeThemeConfig';
